+++ src/main/java/kq/server/bean/User.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public class User {
     String name;
-    int user_id;
+    long user_id;
     int card_left = 10;
     Date last_get_card;
     Date last_get_qian;
@@ -17,11 +17,11 @@ public class User {
     String status;
     int coins;
 
-    public int getUser_id() {
+    public long getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(int user_id) {
+    public void setUser_id(long user_id) {
         this.user_id = user_id;
     }
