+++ src/main/java/kq/server/mapper/UserMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface UserMapper {
     public List<User> getUsers();
-    public User getUser(int user_id);
+    public User getUser(long user_id);
     public void insertUser(User user);
     public void updateUser(User user);
 }
